requirements.txt
# Core dependencies - versiones compatibles
psycopg2-binary==2.9.9
numpy==1.24.3
pgvector==0.2.4

# Sentence transformers - versión más reciente para evitar warnings
sentence-transformers>=2.7.0
transformers>=4.40.0
torch>=2.0.0,<2.2.0
tokenizers>=0.19.0,<0.20.0 
huggingface-hub>=0.19.0,<0.21.0

# Optional
python-dotenv==1.0.0
